Store the CompressOptions struct on the stack
Instead of having an instance variable for CompressOptions it's now put
on the stack. The struct is as of current version only used to indicate
if the buffer passed to compress_update is stable (is there for the next
call to compress_update, or not). flush and compress_end does not use
the struct as of today, so nil can be passed instead.